@@ -42,15 +42,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.0.3",
-    "@angular/cdk": "^18.0.3",
-    "@angular/common": "^18.0.3",
-    "@angular/compiler": "^18.0.3",
-    "@angular/core": "^18.0.3",
-    "@angular/forms": "^18.0.3",
-    "@angular/platform-browser": "^18.0.3",
-    "@angular/platform-browser-dynamic": "^18.0.3",
-    "@angular/router": "^18.0.3",
+    "@angular/animations": "^18.1.0",
+    "@angular/cdk": "^18.1.0",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
     "@dynatrace/dtrum-api-types": "^1.291.1",
     "@hmcts/nodejs-healthcheck": "^1.8.5",
     "@hmcts/properties-volume": "^1.1.0",
@@ -80,14 +80,14 @@
     "@angular-builders/custom-webpack": "^18.0.0",
     "@angular-builders/dev-server": "^7.3.1",
     "@angular-devkit/architect": "^0.1801.0",
-    "@angular-eslint/builder": "^18.0.1",
-    "@angular-eslint/eslint-plugin": "^18.0.1",
-    "@angular-eslint/eslint-plugin-template": "^18.0.1",
-    "@angular-eslint/schematics": "^18.0.1",
-    "@angular-eslint/template-parser": "^18.0.1",
-    "@angular/build": "^18.0.4",
-    "@angular/cli": "^18.0.4",
-    "@angular/compiler-cli": "^18.0.3",
+    "@angular-eslint/builder": "^18.1.0",
+    "@angular-eslint/eslint-plugin": "^18.1.0",
+    "@angular-eslint/eslint-plugin-template": "^18.1.0",
+    "@angular-eslint/schematics": "^18.1.0",
+    "@angular-eslint/template-parser": "^18.1.0",
+    "@angular/build": "^18.1.0",
+    "@angular/cli": "^18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
     "@types/body-parser": "^1.19.5",
     "@types/connect-redis": "^0.0.23",
     "@types/cookie-parser": "^1.4.7",
@@ -99,30 +99,30 @@
     "@types/node": "^20.14.2",
     "@types/nunjucks": "^3.2.6",
     "@types/redis": "^4.0.11",
-    "@typescript-eslint/eslint-plugin": "^7.13.0",
-    "@typescript-eslint/parser": "^7.13.0",
-    "@typescript-eslint/types": "^7.13.0",
+    "@typescript-eslint/eslint-plugin": "^7.16.0",
+    "@typescript-eslint/parser": "^7.16.0",
+    "@typescript-eslint/types": "^7.16.0",
     "axe-core": "^4.9.1",
     "bluebird": "^3.7.2",
     "concurrently": "^8.2.2",
-    "cypress": "^13.11.0",
+    "cypress": "^13.13.0",
     "cypress-axe": "^1.5.0",
-    "ejs": "^3.1.9",
+    "ejs": "^3.1.10",
     "eslint": "^9.5",
     "husky": "^9.0.0",
     "i": "^0.3.7",
     "jest": "^29.7",
     "jest-environment-jsdom": "^29.7",
-    "jest-preset-angular": "^14.1.0",
+    "jest-preset-angular": "^14.1.1",
     "junit-report-merger": "^7.0.0",
     "lodash": "^4.17.21",
     "nodemon": "^3.1.3",
     "npm": "^10.8.1",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "start-server-and-test": "^2.0.4",
     "ts-jest": "^29.1.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
+    "typescript": "^5.5.3",
     "xunit-viewer": "^10.6.1"
   },
   "packageManager": "yarn@3.8.3"
