@@ -1,4 +1,4 @@
-import { HttpClient, HttpClientModule, HttpErrorResponse } from '@angular/common/http';
+import { HttpClient, HttpErrorResponse, provideHttpClient } from '@angular/common/http';
 import { ComponentFixture, TestBed, fakeAsync, flush, tick } from '@angular/core/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
@@ -8,8 +8,10 @@ import { CourthouseData, ErrorMessage, ErrorSummaryEntry } from '@core-types/ind
 import { AppConfigService } from '@services/app-config/app-config.service';
 import { AppInsightsService } from '@services/app-insights/app-insights.service';
 import { CaseService } from '@services/case/case.service';
+import { CourthouseService } from '@services/courthouses/courthouses.service';
 import { ErrorMessageService } from '@services/error/error-message.service';
 import { HeaderService } from '@services/header/header.service';
+import { ScrollService } from '@services/scroll/scroll.service';
 import { of, throwError } from 'rxjs';
 import { CaseSearchResultsComponent } from './case-search-results/case-search-results.component';
 import { SearchComponent } from './search.component';
@@ -24,6 +26,7 @@ describe('SearchComponent', () => {
   let caseService: CaseService;
   let errorMsgService: ErrorMessageService;
   let headerService: HeaderService;
+  let courthouseService: CourthouseService;
   const courts = [
     { courthouse_name: 'Reading', id: 0, created_date_time: 'mock' },
     { courthouse_name: 'Slough', id: 1, created_date_time: 'mock' },
@@ -42,23 +45,20 @@ describe('SearchComponent', () => {
     headerService = new HeaderService();
     errorMsgService = new ErrorMessageService(headerService, mockRouter);
     caseService = new CaseService(httpClientSpy);
+    courthouseService = new CourthouseService(httpClientSpy);
     jest.spyOn(caseService, 'searchCases').mockReturnValue(of([]));
-    jest.spyOn(caseService, 'getCourthouses').mockReturnValue(of(courts));
+    jest.spyOn(courthouseService, 'getCourthouses').mockReturnValue(of(courts));
 
     TestBed.configureTestingModule({
-      imports: [
-        ReactiveFormsModule,
-        FormsModule,
-        HttpClientModule,
-        SearchComponent,
-        CaseSearchResultsComponent,
-        CourthouseComponent,
-      ],
+      imports: [ReactiveFormsModule, FormsModule, SearchComponent, CaseSearchResultsComponent, CourthouseComponent],
       providers: [
         { provide: AppInsightsService, useValue: fakeAppInsightsService },
         { provide: AppConfigService, useValue: fakeAppConfigService },
         { provide: CaseService, useValue: caseService },
+        { provide: CourthouseService, useValue: courthouseService },
         { provide: ErrorMessageService, useValue: errorMsgService },
+        { provide: ScrollService, useValue: { scrollTo: jest.fn() } },
+        provideHttpClient(),
       ],
     });
     fixture = TestBed.createComponent(SearchComponent);
@@ -215,6 +215,31 @@ describe('SearchComponent', () => {
       component.onSubmit();
 
       expect(component.isAdvancedSearch).toEqual(true);
+    });
+
+    it('scroll to top when search button is clicked and fields are filled incorrectly', () => {
+      const scrollService = TestBed.inject(ScrollService);
+      jest.spyOn(scrollService, 'scrollTo');
+
+      component.form.controls['date_to'].setValue('INVALID');
+      component.form.controls['date_from'].setValue('INVALID');
+
+      component.onSubmit();
+
+      expect(scrollService.scrollTo).toHaveBeenCalledWith('app-validation-error-summary');
+    });
+
+    it('scroll to results when search button is clicked and fields are filled correctly', () => {
+      const scrollService = TestBed.inject(ScrollService);
+      jest.spyOn(scrollService, 'scrollTo');
+
+      component.form.controls['case_number'].setValue('1');
+
+      component.form.markAllAsTouched();
+
+      component.onSubmit();
+
+      expect(scrollService.scrollTo).toHaveBeenCalledWith('#results');
     });
   });
 
