@@ -17,6 +17,29 @@ describe('Request audio', () => {
     cy.get('h1').should('contain', 'C20220620001');
   });
 
+  it('should show error for no audio', () => {
+    cy.get('#hearingsTable').should('contain', '1 Dec 2023');
+    cy.get('#hearingsTable a').contains('1 Dec 2023').click();
+    cy.get('.button').contains('Get Audio').click();
+    cy.get('.govuk-error-summary').should('contain', 'There is no audio available for this start time');
+    cy.get('.govuk-error-summary').should('contain', 'There is no audio available for this end time');
+    cy.get('#end-time-error').should('contain', 'There is no audio available for this end time');
+    cy.get('#start-time-error').should('contain', 'There is no audio available for this start time');
+
+    cy.get('#start-time-hour-input').type('08');
+    cy.get('#start-time-minutes-input').type('59');
+    cy.get('#start-time-seconds-input').type('24');
+    cy.get('#end-time-hour-input').type('09');
+    cy.get('#end-time-minutes-input').type('00');
+    cy.get('#end-time-seconds-input').type('24');
+
+    cy.get('.button').contains('Get Audio').click();
+    cy.get('.govuk-error-summary').should('contain', 'There is no audio available for this start time');
+    cy.get('.govuk-error-summary').should('contain', 'There is no audio available for this end time');
+    cy.get('#end-time-error').should('contain', 'There is no audio available for this end time');
+    cy.get('#start-time-error').should('contain', 'There is no audio available for this start time');
+  });
+
   it('should request audio', () => {
     cy.get('#hearingsTable').should('contain', '1 Sep 2023');
     cy.get('#hearingsTable a').contains('1 Sep 2023').click();
@@ -97,17 +120,16 @@ describe('Request audio', () => {
   });
 
   it('should show an error screen if a 409 is returned', () => {
-    cy.get('#hearingsTable').should('contain', '1 Dec 2023');
-    cy.get('#hearingsTable a').contains('1 Dec 2023').click();
+    cy.get('#hearingsTable').should('contain', '5 Jan 2024');
+    cy.get('#hearingsTable a').contains('5 Jan 2024').click();
 
     cy.get('#eventAudioTable .govuk-table__row:nth-child(2) .govuk-checkboxes__item').click();
-    cy.get('#eventAudioTable .govuk-table__row:nth-child(3) .govuk-checkboxes__item').click();
-    cy.get('#start-time-hour-input').should('have.value', '09');
-    cy.get('#start-time-minutes-input').should('have.value', '00');
-    cy.get('#start-time-seconds-input').should('have.value', '00');
-    cy.get('#end-time-hour-input').should('have.value', '09');
-    cy.get('#end-time-minutes-input').should('have.value', '00');
-    cy.get('#end-time-seconds-input').should('have.value', '10');
+    cy.get('#start-time-hour-input').should('have.value', '10');
+    cy.get('#start-time-minutes-input').should('have.value', '32');
+    cy.get('#start-time-seconds-input').should('have.value', '24');
+    cy.get('#end-time-hour-input').should('have.value', '10');
+    cy.get('#end-time-minutes-input').should('have.value', '36');
+    cy.get('#end-time-seconds-input').should('have.value', '24');
 
     cy.get('#download-radio').click({ force: true });
 
@@ -116,9 +138,9 @@ describe('Request audio', () => {
     cy.get('.govuk-grid-column-two-thirds > :nth-child(4)').should('contain', 'C20220620001');
     cy.get('.govuk-grid-column-two-thirds > :nth-child(6)').should('contain', 'Swansea');
     cy.get('.govuk-grid-column-two-thirds > :nth-child(8)').should('contain', 'Defendant Dave');
-    cy.get('.govuk-grid-column-two-thirds > :nth-child(11)').should('contain', '1 Dec 2023');
-    cy.get('.govuk-grid-column-two-thirds > :nth-child(13)').should('contain', '09:00:00');
-    cy.get('.govuk-grid-column-two-thirds > :nth-child(15)').should('contain', '09:00:10');
+    cy.get('.govuk-grid-column-two-thirds > :nth-child(11)').should('contain', '5 Jan 2024');
+    cy.get('.govuk-grid-column-two-thirds > :nth-child(13)').should('contain', '10:32:24');
+    cy.get('.govuk-grid-column-two-thirds > :nth-child(15)').should('contain', '10:36:24');
 
     cy.get('.govuk-button-group > .govuk-button').contains('Confirm').click();
 
