@@ -13,12 +13,12 @@ describe('Request Transcript', () => {
     cy.get('.govuk-table__row').contains('C20220620001');
     cy.get('a').contains('C20220620001').click();
     cy.get('h1').should('contain', 'C20220620001');
-    cy.get('#hearingsTable').should('contain', '1 Sep 2023');
+
+    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
+    cy.get('#hearing-transcripts-tab').click();
   });
 
   it('should get to hearing transcripts tab', () => {
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-caption-l').should('contain', 'Hearing');
     cy.get('.govuk-heading-l').should('contain', '1 Sep 2023');
     cy.get('app-hearing-file').should('contain', 'Courthouse');
@@ -33,8 +33,6 @@ describe('Request Transcript', () => {
   });
 
   it('should get to the request transcript page', () => {
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-caption-l').should('contain', 'Hearing');
     cy.get('.govuk-heading-l').should('contain', '1 Sep 2023');
     cy.get('app-hearing-file').should('contain', 'Courthouse');
@@ -57,8 +55,6 @@ describe('Request Transcript', () => {
   });
 
   it('should return to the hearing transcripts tab', () => {
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-caption-l').should('contain', 'Hearing');
     cy.get('.govuk-heading-l').should('contain', '1 Sep 2023');
     cy.get('app-hearing-file').should('contain', 'Courthouse');
@@ -68,7 +64,6 @@ describe('Request Transcript', () => {
     cy.get('app-hearing-file').should('contain', 'Judge(s)');
     cy.get('app-hearing-file').should('contain', 'HHJ M. Hussain KC');
 
-    cy.get('#transcripts-tab').click();
     cy.get('.flex-space-between > .govuk-heading-m').should('contain', 'Transcripts for this hearing');
     cy.get('.govuk-table__body > :nth-child(1) > :nth-child(1)').should('contain', 'Sentencing remarks');
     cy.get('.govuk-button').should('contain', 'Request a new transcript');
@@ -91,8 +86,6 @@ describe('Request Transcript', () => {
   });
 
   it("should show errors if the form isn't filled in", () => {
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-caption-l').should('contain', 'Hearing');
     cy.get('.govuk-heading-l').should('contain', '1 Sep 2023');
     cy.get('app-hearing-file').should('contain', 'Courthouse');
@@ -117,9 +110,6 @@ describe('Request Transcript', () => {
   });
 
   it('open the request times page if "court log" is selected', () => {
-    // Get to the request transcript page
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
     // Confirm we are in the right place
@@ -135,9 +125,6 @@ describe('Request Transcript', () => {
   });
 
   it('open the request times page if "specified times" is selected', () => {
-    // Get to the request transcript page
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
     // Confirm we are in the right place
@@ -153,10 +140,43 @@ describe('Request Transcript', () => {
     cy.a11y();
   });
 
+  it('show outside times validation error if requested times are outside', () => {
+    cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
+
+    // Confirm we are in the right place
+    cy.get('h1.govuk-heading-l').should('contain', 'Request a new transcript');
+
+    // Fill in the form
+    cy.get('#transcription-type').select('Court log');
+    cy.get('#urgency').select('Overnight');
+    cy.get('.govuk-button').contains('Continue').click();
+
+    // Confirm we are in the right place
+    cy.get('h2.govuk-heading-m').should('contain', 'Events, audio and specific times requests');
+
+    //Fill in times
+    cy.get('#start-hour-input').type('01');
+    cy.get('#start-minutes-input').type('59');
+    cy.get('#start-seconds-input').type('30');
+
+    cy.get('#end-hour-input').type('05');
+    cy.get('#end-minutes-input').type('00');
+    cy.get('#end-seconds-input').type('00');
+
+    cy.get('.govuk-button').contains('Continue').click();
+
+    cy.get('#start-time-error').should(
+      'contain',
+      'Audio not available for timing entered. You must specify a time that matches the audio time available'
+    );
+
+    cy.get('ul.govuk-error-summary__list > li').should(
+      'contain.text',
+      'Audio not available for timing entered. You must specify a time that matches the audio time available'
+    );
+  });
+
   it('open the confirmation page if neither "court log" or "specified times" is selected', () => {
-    // Get to the request transcript page
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
     // Confirm we are in the right place
@@ -172,9 +192,6 @@ describe('Request Transcript', () => {
   });
 
   it('submit a transcript request', () => {
-    // Get to the request transcript page
-    cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
     cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
     // Confirm we are in the right place
@@ -199,9 +216,6 @@ describe('Request Transcript', () => {
 
   describe('Duplicate Transcripts', () => {
     it('should show duplicate error message during normal transcript request', () => {
-      // Get to the request transcript page
-      cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-      cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
       cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
       // Confirm we are in the right place
@@ -224,8 +238,6 @@ describe('Request Transcript', () => {
     });
 
     it('should show duplicate error message during court log transcript request', () => {
-      cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-      cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
       cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
       // Confirm we are in the right place
@@ -240,9 +252,9 @@ describe('Request Transcript', () => {
       cy.get('h2.govuk-heading-m').should('contain', 'Events, audio and specific times requests');
 
       //Fill in times
-      cy.get('#start-hour-input').type('00');
-      cy.get('#start-minutes-input').type('00');
-      cy.get('#start-seconds-input').type('00');
+      cy.get('#start-hour-input').type('03');
+      cy.get('#start-minutes-input').type('33');
+      cy.get('#start-seconds-input').type('33');
 
       cy.get('#end-hour-input').type('05');
       cy.get('#end-minutes-input').type('00');
@@ -260,8 +272,6 @@ describe('Request Transcript', () => {
     });
 
     it('should route to hearings list on duplicate error page', () => {
-      cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-      cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
       cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
       // Confirm we are in the right place
@@ -284,8 +294,6 @@ describe('Request Transcript', () => {
     });
 
     it('should route to specific transcript on duplicate error page', () => {
-      cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-      cy.get(':nth-child(2) > .moj-sub-navigation__link').click();
       cy.get('.govuk-button').should('contain', 'Request a new transcript').click();
 
       // Confirm we are in the right place
