@@ -17,11 +17,12 @@ describe('View Transcript', () => {
 
   it('should get to view transcripts from the case level', () => {
     cy.get('a').contains('C20220620001').click();
-    cy.get('.moj-sub-navigation__link').contains('All Transcripts').click();
+    cy.get('#transcripts-tab').contains('All Transcripts').click();
     cy.get('.govuk-tag--green').should('contain', 'Complete');
     cy.contains('Complete').parents('tr').contains('View').click();
-    cy.contains('Case ID').parents('tr').should('contain', 'C20220620001');
-    cy.get('.govuk-body-m').should('contain', 'Section 4(2) of the Contempt of Court Act 1981');
+    cy.contains('Case ID').parents('.govuk-summary-list__row').should('contain', 'C20220620001');
+    cy.get('.govuk-details__summary-text').click();
+    cy.get('.restriction').should('contain', 'Section 4(2) of the Contempt of Court Act 1981');
   });
 
   it('should get to view transcripts from the hearing level', () => {
@@ -29,7 +30,7 @@ describe('View Transcript', () => {
     cy.get('h1').should('contain', 'C20220620001');
     cy.get('#hearingsTable').should('contain', '1 Sep 2023');
     cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get('.moj-sub-navigation__link').contains('Transcripts').click();
+    cy.get('#hearing-transcripts-tab').contains('Transcripts').click();
     cy.get('.govuk-caption-l').should('contain', 'Hearing');
     cy.get('.flex-space-between > .govuk-heading-m').should('contain', 'Transcripts for this hearing');
     cy.contains('Complete').parents('tr').should('contain', 'Sentencing remarks');
@@ -37,8 +38,8 @@ describe('View Transcript', () => {
     cy.get('.govuk-tag--green').should('contain', 'Complete');
     cy.get('table tr:eq(2) a:contains("View")').click();
     cy.get('h1').should('contain', 'C20220620001_0.docx');
-    cy.contains('Case ID').parents('tr').should('contain', 'C20220620001');
-    cy.get('.govuk-body-m').should('contain', 'Section 4(2) of the Contempt of Court Act 1981');
+    cy.contains('Case ID').parents('.govuk-summary-list__row').should('contain', 'C20220620001');
+    cy.get('.restriction').should('contain', 'Section 4(2) of the Contempt of Court Act 1981');
     cy.a11y();
   });
 
@@ -47,15 +48,15 @@ describe('View Transcript', () => {
     cy.get('h1').should('contain', 'C20220620001');
     cy.get('#hearingsTable').should('contain', '1 Sep 2023');
     cy.get('#hearingsTable a').contains('1 Sep 2023').click();
-    cy.get('.moj-sub-navigation__link').contains('Transcripts').click();
     cy.get('.govuk-caption-l').should('contain', 'Hearing');
+    cy.get('#hearing-transcripts-tab').contains('Transcripts').click();
     cy.get('.flex-space-between > .govuk-heading-m').should('contain', 'Transcripts for this hearing');
     cy.contains('Complete').parents('tr').should('contain', 'Sentencing remarks');
     cy.get('.govuk-button').should('contain', 'Request a new transcript');
     cy.get('.govuk-tag--green').should('contain', 'Complete');
     cy.contains('21 Oct 2023').parents('tr').contains('View').click();
     cy.get('h1').should('contain', 'Document not found');
-    cy.get('.govuk-body-m').should('contain', 'Section 4(2) of the Contempt of Court Act 1981');
+    cy.get('.restriction').should('contain', 'Section 4(2) of the Contempt of Court Act 1981');
   });
 
   it('should get to a completed transcript from Your transcripts', () => {
@@ -78,7 +79,6 @@ describe('View Transcript', () => {
   it('download a transcript', () => {
     cy.get('.moj-primary-navigation__link').contains('Your transcripts').click();
     cy.contains('FGH12345').parents('tr').contains('View').click();
-    // cy.get('.govuk-tag').should('contain', 'Complete');
     cy.get('#download-button').click();
     cy.readFile(path.join(downloadsFolder, 'C20220620001_0.docx')).should('exist');
   });
