@@ -218,6 +218,53 @@ const mockTranscriptionDetails = {
   },
 };
 
+const mockTranscriptionDetailsRejected = {
+  case_id: 1,
+  case_reporting_restrictions: [
+    {
+      hearing_id: 1,
+      event_id: 123,
+      event_name: 'Section 4(2) of the Contempt of Court Act 1981',
+      event_text: '',
+      event_ts: '2023-08-07T09:00:00Z',
+    },
+  ],
+  case_number: 'C20220620001',
+  courthouse: 'Swansea',
+  courtroom: '5',
+  courthouse_id: 1,
+  status: 'Rejected',
+  from: 'MoJ CH Swansea',
+  requestor: {
+    user_id: 1,
+    user_full_name: 'Joe Smith',
+  },
+  received: '2023-11-17T12:53:07.468Z',
+  requestor_comments: 'Please expedite my request',
+  rejection_reason: 'This request will take longer to transcribe within the urgency level you require.',
+  defendants: ['Defendant Dave', 'Defendant Bob'],
+  judges: ['HHJ M. Hussain KC	', 'Ray Bob'],
+  transcript_file_name: 'C20220620001_0.docx',
+  hearing_date: '2023-11-08',
+  transcription_urgency: {
+    transcription_urgency_id: 1,
+    description: 'Standard',
+    priority_order: 4,
+  },
+  request_type: 'Specified Times',
+  request_id: 123456789,
+  transcription_id: 1,
+  transcription_start_ts: '2023-06-26T13:00:00Z',
+  transcription_end_ts: '2023-06-26T16:00:00Z',
+  transcription_object_id: 109,
+  is_manual: true,
+  hearing_id: 1,
+  requestor: {
+    user_id: 1,
+    user_full_name: 'Eric Bristow',
+  },
+};
+
 const mockTranscriptionDetailsTwo = {
   case_id: 2,
   case_reporting_restrictions: [
@@ -288,6 +335,7 @@ const mockTranscriptionDetailsNoName = {
     },
   ],
 };
+
 const defaultUnassignedTranscriptions = [
   {
     transcription_id: 4,
@@ -564,7 +612,16 @@ router.get('/:transcriptId', (req, res) => {
       res.status(200).send(mockTranscriptionDetailsNoName);
       break;
     case '4':
-      res.status(200).send(mockTranscriptionDetails);
+      res.status(200).send(mockTranscriptionDetailsRejected);
+      break;
+    case '5':
+      res.status(200).send({
+        ...mockTranscriptionDetails,
+        transcription_start_ts: null,
+        transcription_end_ts: null,
+        request_type: 'Sentencing Remarks',
+        status: 'Complete',
+      });
       break;
     default:
       res.status(200).send(mockTranscriptionDetailsTwo);
@@ -572,9 +629,12 @@ router.get('/:transcriptId', (req, res) => {
 });
 
 router.patch('/:transcriptId/document', (req, res) => {
-  assignedTranscriptions = assignedTranscriptions.map((t) => {
+  let assignedTranscriptionsCopy = [...defaultAssignedTranscriptions];
+
+  assignedTranscriptionsCopy = assignedTranscriptionsCopy.map((t) => {
     if (t.transcription_id == req.params.transcriptId) {
-      t.status = 'Complete';
+      const updatedTranscription = { ...t, status: 'Complete' };
+      return updatedTranscription;
     }
     return t;
   });
@@ -582,9 +642,12 @@ router.patch('/:transcriptId/document', (req, res) => {
 });
 
 router.post('/:transcriptId/document', (req, res) => {
-  assignedTranscriptions = assignedTranscriptions.map((t) => {
+  let assignedTranscriptionsCopy = [...defaultAssignedTranscriptions];
+
+  assignedTranscriptionsCopy = assignedTranscriptionsCopy.map((t) => {
     if (t.transcription_id == req.params.transcriptId) {
-      t.status = 'Complete';
+      const updatedTranscription = { ...t, status: 'Complete' };
+      return updatedTranscription;
     }
     return t;
   });
@@ -619,7 +682,7 @@ router.patch('/:transcriptId', (req, res) => {
 
 router.post('/', (req, res) => {
   //If start time is below then return 409
-  const exists = req.body.start_date_time.indexOf('00:00:00Z') !== -1 && true;
+  const exists = req.body.start_date_time.indexOf('03:33:33Z') !== -1 && true;
   const dupe = req.body.transcription_type_id == 0 && true;
 
   if (exists || dupe) {
