@@ -1,18 +1,23 @@
 import { Transcription, TranscriptionDocumentSearchResult, TranscriptionStatus } from '@admin-types/transcription';
 import { DatePipe } from '@angular/common';
-import { ComponentFixture, TestBed, fakeAsync } from '@angular/core/testing';
+import { signal } from '@angular/core';
+import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { ActivatedRoute, Router } from '@angular/router';
 import { CourthouseData } from '@core-types/index';
 import { LuxonDatePipe } from '@pipes/luxon-date.pipe';
 import { CourthouseService } from '@services/courthouses/courthouses.service';
-import { TranscriptionAdminService } from '@services/transcription-admin/transcription-admin.service';
+import {
+  defaultFormValues,
+  TranscriptionAdminService,
+} from '@services/transcription-admin/transcription-admin.service';
 import { DateTime } from 'luxon';
-import { of } from 'rxjs';
+import { BehaviorSubject, of } from 'rxjs';
 import { TranscriptsComponent } from './transcripts.component';
 
 describe('TranscriptsComponent', () => {
   let component: TranscriptsComponent;
   let fixture: ComponentFixture<TranscriptsComponent>;
+  let router: Router;
 
   const MOCK_SEARCH_RESULT = [
     { id: 1, courthouse: { id: 1 }, status: { id: 1 } },
@@ -103,6 +108,12 @@ describe('TranscriptsComponent', () => {
             mapResults: jest.fn().mockReturnValue(MOCK_MAPPING),
             search: jest.fn().mockReturnValue(of(MOCK_SEARCH_RESULT)),
             searchCompletedTranscriptions: jest.fn().mockReturnValue(of(MOCK_COMPLETED_SEARCH_RESULT)),
+            searchFormValues: signal(defaultFormValues),
+            searchResults: signal([]),
+            completedSearchResults: signal([]),
+            tab: signal('Requests'),
+            isAdvancedSearch: signal(false),
+            hasSearchFormBeenSubmitted$: new BehaviorSubject<boolean>(false),
           },
         },
         {
@@ -119,6 +130,8 @@ describe('TranscriptsComponent', () => {
 
     fixture = TestBed.createComponent(TranscriptsComponent);
     component = fixture.componentInstance;
+    router = TestBed.inject(Router);
+    jest.spyOn(router, 'navigate').mockResolvedValue(true);
     fixture.detectChanges();
   });
 
@@ -127,16 +140,16 @@ describe('TranscriptsComponent', () => {
   });
 
   it('should start loading when search is triggered', () => {
-    jest.spyOn(component, 'startLoading');
+    component.loadingResults.set = jest.fn();
     component.search$.next({});
-    expect(component.startLoading).toHaveBeenCalled();
+    expect(component.loadingResults.set).toHaveBeenCalledWith(true);
   });
 
   it('should stop loading after search is completed', () => {
-    jest.spyOn(component, 'stopLoading');
+    component.loadingResults.set = jest.fn();
     jest.spyOn(component.transcriptService, 'search').mockReturnValue(of([]));
     component.search$.next({});
-    expect(component.stopLoading).toHaveBeenCalled();
+    expect(component.loadingResults.set).toHaveBeenCalledWith(false);
   });
 
   it('should call search when search is triggered', () => {
@@ -197,6 +210,7 @@ describe('TranscriptsComponent', () => {
     component.errors = [{ fieldId: '1', message: 'error' }];
     component.clearSearch();
     expect(component.errors).toEqual([]);
+    fakeAsync;
   });
 
   it('should navigate to the transcript page if only one result is returned', () => {
@@ -207,12 +221,37 @@ describe('TranscriptsComponent', () => {
     expect(component.router.navigate).toHaveBeenCalledWith(['/admin/transcripts', 1]);
   });
 
+  it('only search for transcript requests if the tab is "Requests"', fakeAsync(() => {
+    jest.spyOn(component.transcriptService, 'search');
+    component.tab.set('Requests');
+    component.isSubmitted$.next(true);
+    component.search$.next({});
+
+    tick();
+
+    expect(component.transcriptService.search).toHaveBeenCalled();
+    expect(component.transcriptService.searchCompletedTranscriptions).not.toHaveBeenCalled();
+  }));
+
+  it('only search for completed transcripts if the tab is "Completed transcripts"', fakeAsync(() => {
+    jest.spyOn(component.transcriptService, 'searchCompletedTranscriptions').mockReturnValue(of([]));
+    component.tab.set('Completed transcripts');
+    component.isSubmitted$.next(true);
+    component.search$.next({});
+
+    tick();
+
+    expect(component.transcriptService.searchCompletedTranscriptions).toHaveBeenCalled();
+    expect(component.transcriptService.search).not.toHaveBeenCalled();
+  }));
+
   describe('completedResults$', () => {
     it('should return the results when the search is completed', () => {
       let result;
 
       component.completedResults$.subscribe((results) => (result = results));
 
+      component.tab.set('Completed transcripts');
       component.search$.next({});
       component.isSubmitted$.next(true);
 
@@ -229,6 +268,7 @@ describe('TranscriptsComponent', () => {
 
       component.completedResults$.subscribe();
 
+      component.tab.set('Completed transcripts');
       component.search$.next({});
       component.isSubmitted$.next(true);
 
