@@ -1,6 +1,7 @@
 import { TranscriptionSearchFormValues } from '@admin-types/transcription';
 import { AsyncPipe } from '@angular/common';
-import { Component, inject } from '@angular/core';
+import { Component, inject, signal } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { Router } from '@angular/router';
 import { GovukHeadingComponent } from '@common/govuk-heading/govuk-heading.component';
 import { LoadingComponent } from '@common/loading/loading.component';
@@ -8,8 +9,11 @@ import { TabsComponent } from '@common/tabs/tabs.component';
 import { ValidationErrorSummaryComponent } from '@common/validation-error-summary/validation-error-summary.component';
 import { TabDirective } from '@directives/tab.directive';
 import { CourthouseService } from '@services/courthouses/courthouses.service';
-import { TranscriptionAdminService } from '@services/transcription-admin/transcription-admin.service';
-import { BehaviorSubject, Subject, combineLatest, map, of, shareReplay, switchMap, tap } from 'rxjs';
+import {
+  defaultFormValues,
+  TranscriptionAdminService,
+} from '@services/transcription-admin/transcription-admin.service';
+import { combineLatest, map, of, shareReplay, Subject, switchMap, tap } from 'rxjs';
 import { SearchCompletedTranscriptsResultsComponent } from './search-completed-transcripts-results/search-completed-transcripts-results.component';
 import { SearchTranscriptsFormComponent } from './search-transcripts-form/search-transcripts-form.component';
 import { SearchTranscriptsResultsComponent } from './search-transcripts-results/search-transcripts-results.component';
@@ -41,13 +45,18 @@ export class TranscriptsComponent {
   transcriptionStatuses$ = this.transcriptService.getTranscriptionStatuses().pipe(shareReplay(1));
 
   search$ = new Subject<TranscriptionSearchFormValues | null>();
-  loading$ = new Subject<boolean>();
-  isSubmitted$ = new BehaviorSubject<boolean>(false);
+  isSubmitted$ = this.transcriptService.hasSearchFormBeenSubmitted$;
+  tab = this.transcriptService.tab;
+
+  loadingResults = signal(false);
+  loadingCompletedResults = signal(false);
 
   results$ = combineLatest([this.search$, this.isSubmitted$, this.courthouses$, this.transcriptionStatuses$]).pipe(
-    tap(() => this.startLoading()),
+    tap(() => {
+      this.loadingResults.set(true);
+    }),
     switchMap(([values, isSubmitted, courthouses, statuses]) => {
-      if (!values || !isSubmitted) {
+      if (!values || !isSubmitted || this.tab() === 'Completed transcripts') {
         return of(null);
       }
       return (
@@ -57,8 +66,9 @@ export class TranscriptsComponent {
           .pipe(map((results) => this.transcriptService.mapResults(results, courthouses, statuses)))
       );
     }),
+    takeUntilDestroyed(),
     tap((results) => {
-      this.stopLoading();
+      this.loadingResults.set(false);
       if (results?.length === 1) {
         //navigate to the transcript details page
         this.router.navigate(['/admin/transcripts', results[0].id]);
@@ -67,15 +77,18 @@ export class TranscriptsComponent {
   );
 
   completedResults$ = combineLatest([this.search$, this.isSubmitted$]).pipe(
-    tap(() => this.startLoading()),
+    tap(() => {
+      this.loadingCompletedResults.set(true);
+    }),
     switchMap(([values, isSubmitted]) => {
-      if (!values || !isSubmitted) {
+      if (!values || !isSubmitted || this.tab() === 'Requests') {
         return of(null);
       }
       return this.transcriptService.searchCompletedTranscriptions(values);
     }),
-    tap(() => this.stopLoading()),
+    takeUntilDestroyed(),
     tap((results) => {
+      this.loadingCompletedResults.set(false);
       if (results?.length === 1) {
         // navigate to the transcript document details page
         this.router.navigate(['/admin/transcripts/document', results[0].transcriptionDocumentId]);
@@ -83,24 +96,33 @@ export class TranscriptsComponent {
     })
   );
 
-  startLoading() {
-    this.loading$.next(true);
-  }
-
-  stopLoading() {
-    this.loading$.next(false);
+  constructor() {
+    this.results$.subscribe((results) => results && this.transcriptService.searchResults.set(results));
+    this.completedResults$.subscribe(
+      (results) => results && this.transcriptService.completedSearchResults.set(results)
+    );
   }
 
   onSearch(values: TranscriptionSearchFormValues) {
-    if (this.isSubmitted$.value === false) {
+    if (this.isSubmitted$.getValue() === false) {
       this.isSubmitted$.next(true);
     }
+    this.transcriptService.searchFormValues.set(values); // save form state
     this.search$.next(values); // Trigger the search
   }
 
   clearSearch() {
-    this.isSubmitted$.next(false);
+    this.transcriptService.searchFormValues.set(defaultFormValues);
+    this.transcriptService.searchResults.set(null);
+    this.transcriptService.isAdvancedSearch.set(false);
+    this.transcriptService.completedSearchResults.set(null);
+    this.transcriptService.hasSearchFormBeenSubmitted$.next(false);
     this.search$.next(null);
     this.errors = [];
+  }
+
+  onTabChanged(tab: string) {
+    this.clearSearch();
+    this.transcriptService.tab.set(tab);
   }
 }
