@@ -6,6 +6,7 @@ import { BreadcrumbComponent } from '@common/breadcrumb/breadcrumb.component';
 import { DataTableComponent } from '@common/data-table/data-table.component';
 import { GovukBannerComponent } from '@common/govuk-banner/govuk-banner.component';
 import { GovukHeadingComponent } from '@common/govuk-heading/govuk-heading.component';
+import { LoadingComponent } from '@common/loading/loading.component';
 import { BreadcrumbDirective } from '@directives/breadcrumb.directive';
 import { TableRowTemplateDirective } from '@directives/table-row-template.directive';
 import { BytesPipe } from '@pipes/bytes.pipe';
@@ -20,6 +21,8 @@ import { AssociatedAudioTableComponent } from '../associated-audio-table/associa
 @Component({
   selector: 'app-view-transformed-media',
   standalone: true,
+  templateUrl: './view-transformed-media.component.html',
+  styleUrl: './view-transformed-media.component.scss',
   imports: [
     GovukHeadingComponent,
     BreadcrumbComponent,
@@ -35,9 +38,8 @@ import { AssociatedAudioTableComponent } from '../associated-audio-table/associa
     RouterLink,
     GovukBannerComponent,
     AssociatedAudioTableComponent,
+    LoadingComponent,
   ],
-  templateUrl: './view-transformed-media.component.html',
-  styleUrl: './view-transformed-media.component.scss',
 })
 export class ViewTransformedMediaComponent {
   transformedMediaService = inject(TransformedMediaService);
